chore: migrate benchmark package to apps directory
Move benchmark code and configuration to apps/matrixcurator-benchmark, update imports, and improve dataset synchronization performance with async parsing and rate limiting.
