+++ apps/matrixcurator-benchmark/src/matrixcurator_benchmark/core/__init__.py
@@ -1,0 +1,13 @@
+# src/benchmark/core/__init__.py
+"""Core module for benchmark execution."""
+from matrixcurator_benchmark.core.decorators import benchmark, fixture, parametrize
+from matrixcurator_benchmark.core.exceptions import FailBenchmark, SkipBenchmark
+
+__all__ = [
+    "FailBenchmark",
+    "SkipBenchmark",
+    "benchmark",
+    "fixture",
+    "parametrize",
+]
+
