--- src/benchmark/core/__init__.py
@@ -1,13 +1,0 @@
-# src/benchmark/core/__init__.py
-"""Core module for benchmark execution."""
-from src.benchmark.core.decorators import benchmark, fixture, parametrize
-from src.benchmark.core.exceptions import FailBenchmark, SkipBenchmark
-
-__all__ = [
-    "FailBenchmark",
-    "SkipBenchmark",
-    "benchmark",
-    "fixture",
-    "parametrize",
-]
-
